refactor: extract independent addTasks
